Separate preprocessing for blob detection and watershed
Watershed has been using the same preprocessed ROI as for blob detection, and this change uses the raw ROI for separate preprocessing step for each operation. The main side-effect is that preprocessing and segmentation are integrated during automated processing, meaning that the ROI size is the same for each step rather than allowing the preprocessing ROI to be smaller than segmentation as previously.

- Further modularize saturation and denoising
- Integrate preprocessing into segmentation step
- Save a separate processed ROI for viewing