Fix xz orientation to use aspect ratio based on scaling and flipped image
Previously the aspect ratio for the overview image for xz orientation was based the image size, which did not depend on image scaling but rather the number of sections imaged, even though this aspect ratio had worked perhaps better for previous images. This fix uses scaling instead. The overview image is now also flipped when in xz orientation since our microscopy has typically imaged the ventral surface of the brain first. We may eventually want to flip all overview images or create a separate option.