Option to save ROI from original image file
Saving a small region of the original file can be helpful for analyzing images on different computers without having to transfer the entire stack. Previously the image5d_proc file held the processed image file, after denoising and filtering. Current versions have no longer saved this file since they take a large amount of space, and on-the-fly processing showed a similar view. This update repurposes image5d_proc to save a small section of the original, unprocessed image, matching the region segmented and saved in info_proc. These two files along with the image5d_info are now sufficient to reload the processed file for analyzing.

- "--saveroi" command-line option to flag saving the original image in the ROI processed for segmentation
- Option to decline importing the original, non-Numpy image file when the Numpy file cannot be read
- Load the image5d info file before attempting to load the image in case the full image does not exist, in which case image5d_proc will be used instead if in proc loading mode
- When showing filtered image, no longer tries to extract from image5d_proc