Reintroduce output directory flag for stitching
Previously the output directory flag was passed to the Stitching plugin, but it requires the configuration file to be in this directory as well. To minimize changes to usage of the plugin, this flag is now used in the stitching Bash script to manage the directory including moving fused image files into it.

- "-o" flag to set output directory
- Make output directory and move fused images files into it
- Assume fused files are named "img_t*", in the same location as the input image file