Separate local image parent path from S3 destination and add destination command-line flag
The image parent path had been used for the S3 destination path, and the local destination was hard-coded to start with "/data". This update removes the parent path from the S3 destination path to use the set S3 path directly instead and removes the hard-coded destination, defaulting instead to the local path unless set from a new command-line flag.