Draggable circles to reposition blob locations
Previously relocating blobs required marking them as incorrect and adding a new segment in the correct place. This update uses a new DraggableCircles class to allow dragging blob circles into the correct position. The class also stores segment information to avoid mapping to circle collections and storing new circles in dictionaries. Note that circles can not be dragged to different plots to move z-planes, for example.

- DraggableCircles class to store segment information and handle picking and dragging
- Replace PatchCollection for segments in the given z-plane with individual DraggableCircles
- Click order changed when picking circles already present
- Skips red flag for newly added circles