Show blobs from border region when reloading ROI
Previously blobs from the padding region but not the border region were shown when displaying an ROI and only with processed blobs loaded, or no outside blobs without loaded blobs. This update simplifies and improves handling of the various blob display scenarios, including display of border blobs with or without loaded blobs.

- Fix missing border blobs when displaying a saved ROI along with processed blobs
- Change blob coordinate handling to start with relative coordinates for consistency and simplification, with absolute coordinates added later
- Generalize building visualizer segments array