Save the last removed circle for pasting in case the user did not intend to remove it
User added circles are removed when their color is changed from yellow to off. When the user accidentally picks a circle, changing its color, the original color cannot be restored without removing this circle and adding a new one. This update saves the last removed circle so that it can be readded using the paste function ("v"+click) onto the same plot, which creates a new circle in the same place and with an empty color.