Make processed file output naming consistent with non-processed files
Non-processed files (eg raw image Numpy files) have an underscore before the series designation in filenames. Processed files previously omitted this underscore to allow additional characters such as the offset and size values to be appended directly to the end of the original image's filename, following the extension without separation. This method assumed that the extension were known (eg ".czi") or only 3 characters so that the following characters could be extracted. This update adds an underscore to processed files so that they match the style of non-processed files. This change now requires the extra characters to be placed within the image filename string but allow for extensions of any size or value.

- Add underscore after portion spliced into processed filename
- Allow arbitrary extensions for processed file loading
- Clarify feedback when ROI not loaded
- Removed conversion code for old monolithic saved image and info format now that most files have been converted or should be made anew
- Revert sample script variables to previous ones to set while extracting basename and extension programatically