Append experiment code to BigStitcher dataset output names
Assume that the input image filename has an experiment code separated by a dash at the start of its name. If not, the full name of the image will be used. This update ensure that the output names are unique among experiments and avoids overwriting output images generated in the same directory.