Label mirroring also adds lateral edges
The Allen Brain Atlas reference labels for the developing mouse brain does not include the lateral edges of the sample in addition to the missing hemisphere. This update builds these edges by finding the corresponding area of the reference atlas and resizing the image in the next plane of the labels atlas to fit this area.

- Build lateral edges by extending the last labels edge plane
- Adjustable min/max for markers in Random-walker segmentation, with max adjusted to >=