No longer give default segments row
Previously needed default segments row to avoid a crash, presumably from TraitsUI. This default row could cause problems, however, since it could be saved and give false information. As of the current TraitsUI used (updated from 5.1.0 to 5.2.0pre), this default no longer appears to be needed, so we won't included it anymore.