Remove BigStitcher script bounding box setting
Setting the bounding box to "All Views" without dimensions as used for the patched Multiview-Reconstruction used in BigStitcher leads to an error now that the unpatched, official version is used. Multiview-Reconstruction now appears to fuse all views by default, however, so the setting can be simply removed. The selection is reported as "Currently Selected Views", but the bounding box displayed covers the entire area.