No segment filtering if not showing circles
Avoid segment filtering during 2D plot display when not showing circles since the segments will not be shown. This change also serves to prevent a crash when the user has added a segment in circle mode, then opens another plot without circles, which attempts to load the inside ROI mask but encounters an index error.