Measure volumes corresponding to labels where signal is actually present (currently turned off)
Previously volume measurements for each labeled region is calculated based on the labels image itself under the assumption that the labels image would have been registered and deformed to find the shape of the experiment image. This update only includes the portion of each corresponding region where sufficient signal is present. This measurement takes considerable time as much of the large, original image must be evaluated.

- Volume measurement by signal in corresponding region (turned off by default)
- Signal threshold value as a constant