Channel changer also resets GUI state
Previously the user could restore an ROI, select a new channel, and display the blobs from the old channel. This update resets the GUI state with channel changes to ensure that the ROI and blobs will need to be reloaded. Future updates may defer blob filtering by channel until a subsequent step (eg 2D plot) to avoid requiring a reset.