Allow overview image scrolling even when mouse is outside of overview image
Overview image scrolling had been restricted to scrolling only when mouse was within an overview image to allow scrolling other images or individual overview images. Now that overview image is synchronized and other images are not scrolled, this update simplifies scrolling by not requiring the mouse to be within an overview image. If zoomed plot images will be scrolled in the future, scrolling behavior could be tailored to individual image types.