Workaround for floating point exception when resizing 1px thick image
Resizing 1px thick image with "reflect" mode causes a floating point exception in some cases (reliably for some shapes but not for others for some reason) as reported in https://github.com/scikit-image/scikit-image/issues/3001. As a workaround, this update simply changes the mode to "edge" as mentioned in that bug report for images 1px thick in any dimension.