Workaround to prevent run script updating from adjusting other image variables
The sed command in the run script setup tool adjusts "IMG" as well as "TIFF_IMG" variables when it should only update the former. The usual command to update only the first occurrence does not work on non-GNU sed, such as that on the Mac. This update has a workaround that includes the initial part of the default path for the IMG variable to replace.