Quit ImageJ/Fiji from script only if wrote fused file
Turning off stitched file writing implies that the user wishes to review the image before writing. This update takes advantage of how leaving ImageJ/Fiji open causes the enclosing script to pause, allowing the user to review the image and manually close the app, which resumes the enclosing script including running the stitch script with writing now on.

- Only close ImageJ/Fiji after writing stitched image
- Remove unnecessary end of line semicolons