Approximate the middle of a region based on the middle value of its list of coordinates
The centroid of a bounding box for a given region does not always appear to be in the region itself in cases of irregularly shaped or sparsely scattered regions. get_region_middle ensures that the returned coordinate exists within the region itself by taking a list of the region coordinates and returning a value near its middle.

- Find the approximately middle coordinate within a given region
- Fix a rounding error when scaling coordinates to get labels