Adjust position of 3D blobs for isotropy rather than stretching them
Stretching blobs for isotropy places them in the correct position but makes them appear elongated. This update scales the position of blobs instead. A side-effect is that a copy of the segments must be made beforehand to avoid adjusting the master list of segments, which could have a performance impact when displaying many blobs at once.