Allow saving ROIs with missing verifications but continue to warn
Frequently ROIs need to be saved as drafts to check whether the blobs have been saved properly or when shutting down the app before completing a session. This behavior is now allowed, though a strong warning is also supplied to encourage the user to check for missing verifications at the given coordinates.