Remove catch-all level for volume grouping
Catch-all level was used for volumes that did not fit the target level but ended up capturing volumes that did not have a parent at the given level as well as any volume from a higher level. This fix removes the catch-all grouping and improves efficiency by checking only those regions below (numerically higher than) the given level as well as stopping for a volume as soon as the level of a checked parent is above (numerically lower than) the target. Now that some levels have no regions, the bar plotter will check for empty lists.