Generalize volume plots to accept arbitrary measurements
The volume plotter has been hard-coded to plot volume and density to measurements. With the need to plot additional measurements such as absolute blob counts, this update generalizes the plotter to plot an arbitrary number of measurements. Bar stats are stored in a dictionary rather than multiple lists, and each measurement is plotted in a separate sub-plot. Stats outputs measurement units to specify the y-axis unit display.