Curate labels image when creating a new atlas by in-painting missing pixels and carving out extraneous ones
When creating a new atlas, curate the resulting labels image based on the fixed image by in-painting pixels that are present in the fixed image but missing from the labels image, using nearest neighbors from within the labels image to preseve the exact set of labels. Also, carve out pixels from the labels image that are below threshold in the fixed image.