Output group registered image to separate filename and new atlas to atlas filename
Revert to outputting group registered image to "grouped.mhd" to distinguish it from the atlas. Once a new atlas is generated, output it to "atlasVolume.mhd". This atlas will likely be the group registered file but allows flexibility in case the atlas is not. The original atlas file will be named "atlasVolumeTemplate.mhd" to signify that it was from the original template but not in its original position.