Allow configuring the extent of mirroring
Change the "labels_mirror" register setting from boolean to a tuple specifying the extent of the labels image to mirror in case the labels image does not contain exactly one half of the image. Also fix mirroring from the start of the labels image, which previously did not include any planes from the image.