Downsample image to specific target size
Downsampling has supported rescaling by a given factor, which preserves aspect ratio but does not allow selectively preserving information in a given dimension. A target size can now be specified as a register profile setting to target a given final image size such as that from an atlas to which the image will be registered.

- "target_size" register profile setting to specify desired final image output size during Numpy image array transposition and reloading during volume/density calculation
- "resize_factor" can be set to None to avoid resizing for images that are already have near identical sizes
- "_abae18pt5" register profile modifier with specific size for the ABA E18pt5 atlas
- "test" register profile for simple testing
- Fix to adjust scaling after rescaling during registration image transposition
- Workaround to fix spacing slightly mismatch after applying transformix under some (as of yet ill-defined) situations