Erase rather than crop low x region during groupwise registration
Option to erase instead of cropping planes outside a given region, similar to truncation function but based on empty area. Apply this option to groupwise registration to maintain the shape of the image, which is necessary when the input image was resized to a given target size. When this input does not have to match a given target, erasing rather than cropping should still have the advantage of leaving the image's subject toward to the center of the image dimensions.