Option to extend border regions of groupwise registered image with corresponding regions from first image
Groupwise registration crops the variable region of images to avoid the registration from warping images lacking those regions into them. The side effect is that the groupwise registered image completely lacks this region and will be warped into it when acting as the moving image to registered to another image. As a workaround, "extend_borders" option specifies the region to re-incorporate from the raw first input image for groupwise registration. This image is assumed to be close in alignment the groupwise image since the first image serves as the template for output image dimensions.