Allow grid spacing schedule with assignment per dimension
Registration setup simplifies resolution settings by assuming that the number of values in the spacing schedule equals the number of resolutions. To allow varying values by dimension for each resolution, check whether any duplicate resolutions are found as a proxy for this case. Note that schedules with completely unique values even by dimension will not be identified correctly, though this case is likely rare since most images are roughtly isotropic in 2 dimensions.