Fix image flip during high speed zooming
When the zoom speed is fast, the image may flip since the low and high values for x or y limits may cross one another. Check to ensure that the order of these values is the same in the updated limits as in the current limits since "high" is not necessarily higher than "low" depending on the image origin.