Order label smoothing by region size
Label smoothing is currently dependent on the order of labels. The order has been determined simply by the ID number of the label, which roughly corresponds to hierarchically broader to finer structures but is dependent on label numbering implementation and in the case of ABA follows each branch in sequence. Rather than following label ID, order label smoothing by size, working from largest to smallest.