Reduce memory use for colormaps of labels with large ranges
Discrete colormaps have been generated by making a separate color for each integer within the range of labels, but many colors are unused if the labels have gaps. In the caase of same atlases, the gaps are very large, requiring an excessive use of memory to generate a large colormap. To reduce memory usage and simplify discrete colormap generation, use a BoundaryNorm with the sorted, unique labels themselves as the boundaries to allow gaps while generating colors only for each unique label. Since this colormap is closely linked with ListedColormap and carries a requires normalization object, create DiscreteColormap as a new class extending ListedColormap.