Smoothing metric based on broad volume size reduction
Propose a metric to measure degree of appropriate smoothing, defined as smoothing that retains the general shape and placement of the region. This metric compares the difference in size of a broad, smooth volume encompassing ragged edges before and after the smoothing algorithm, giving a measure of reduction in size and thus increase in compactness by smoothing. To penalize inappropriate smoothing, the smoothed volume lying outside of the original broad volume is subtracted from this size reduction, accounting for unwanted deformation that brings the region outside of its original bounds. This metric makes use of the assumption that any change in one region will affect neighboring regions, so overly compacting one region will cause neighboring regions to become overly deformed to fill in the missing space left behind by the excessive compaction. Optimizing this metric seeks to balance compaction with retaining the original placement and overall shape of each region.

- Label smoothing metric
- Save borders image