Testing smoothing metric adjustment for full compaction measurement
Use the full smoothed volume for reduction fraction term as a full measure of compaction. Compaction measurements will be less and more reflective of the true amount of compaction. The volume outside the original area will be doubly counted against the metric since this volume will be included in the penalty term to capture a measure of displacement, but this approach obviates the need for a penalty term weighting since it effectively doubles the penalty term. The penalty weighting is left in case it may be used to dial down the displacement penalty.