Smoothing metric by surface area
Add an additional smoothing metric option for measurement of surface area change. Smoothing would be expected to decrease the surface area of labels by reducing the amount of interleaving among labels. Displacement is taken as the departure from the smoothed perimeter from the original perimeter, measured as the distance transform of the pixels corresponding to one perimeter to the other perimeter.

- Option for smoothing metric by surface area
- Generalized method to get the perimeter region of an array of arbitrary dimensions by subtracting the eroded region
- Generalized method to get the distances between two borders