Option to show distances in the areas where compaction may have been appropriate
A caveat to the displacement measurement for smoothing metric by surface area is borders by definition must shift slightly to achieve compaction, so part of the displacement term will compise appropriate shifts. To measure the amount of distance in this potential space, take a closing filter around the border itself, which tends to leave straight lines alone but fills in areas that may be compacted. Use the filter_size parameter to flag this border distance calculation, defaulting to None since otherwise filters are not needed for smoothing metric by surface area. In that case, padding for the each label region is also decreased to a default value of 2 since less padding (but apparently not none from testing) is necessary for surface area calculations.