Change borders distance measurement option to measure distance from filled border
Giving a filter size to the surface area smoothing metric generated a statistic for the distance within a filled border region to the original border, and this update changes this functionality by applying it to the distance measurements themselves. When applied, distances from the smoothed border will be measured in relationship to the filled border to avoid penalizing displacements that could have been from appropriate compaction, such as ragged or sharp border regions. Also simplify borders_distance to always return a tuple.