Signed distance transform implementation
Implement a signed version of distance transform by providing a mask in which all values are negative. This version targets the case where values within a border, defined as those in the full image, are negative, whereas those outside are positive. The direct distances smoothing metric makes use of this signage by summing displacements only on positive values.