Recursively use equal or smaller regions while extending label planes
Recursively adjust the size of the labels and atlas image regions used to extend planes, which both improves performance and minimizes the chance of picking up artifacts in the rest of each plane. This change assumes that the signal is getting progressively smaller or at least remains the same size while extending the labels signa to the near edge.