Option to duplicate planes while mirroring
Option during mirroring to set a starting fraction at which planes are duplicated from the preceding plane or range of preceding planes. This option may be useful when the true midline does not correspond to the labels midline, such as when labels extend far past the midline and would be lost if mirrored at the true midline.