Fix setting background in discrete colormap when a label one above background exists
The background does not appear to map correctly when the set of labels contains a value one above that of the background, presumably because another boundary condition arises where the background is just below this value because of the offset of 1 used to avoid off-by-one errors. To ensure that the label is somewhere in the middle of the corresponding boundary, use an offset of 0.5 instead.