Display full super-region when searching for region by ID
Currently searching for a region by ID simply shifts the ROI to the center of the region, using the current ROI size, without fitting the ROI to the region. Provide an option to show the region only corresponding to the label itself, including any sub-labels, when searching for a region with "Raw" unchecked.

- Option to show full 3D label when searching for a region
- Accept sequence of IDs when getting bounding box for label