Do not display orientation axes after showing 3D label
Occasionally the displayed 3D surface does not save properly for unclear reasons, and the object cannot be saved manually since the orientation axes are also in the scene. As a workaround, defer showing orientation axes when showing 3D labels to allow the user to manually save the scene and also to show the axes manually if preferred. Also fix doc about processing when preparing ROI.