Change P56 mouse atlas names to avoid conflict
ABA P56 atlas profiles conflicted because one name was contained in the other. As a workaround to avoid this conflict while allowing simple matching, change the ABA P56 profile modifier names. This change also has the benefit of clarifying that the P56 developing atlas is part of the normal series and highlighting the other P56 atlas as "adult".