Enforce exact match for modifier profile names
Modifier names have been allowed to match as long as the name was contained within the full settings string, which allowed backward compatibility with modifiers that were previously longer but now truncated (eg "lightsheet5x" to "lightsheet"). While this behavior was more flexible, it caused confusion when modifier names simply appended characters to other names.

To fix this behavior, use the separator term when adding modifiers to check whether the modifier exists as a distinct entity, within surrounding separators, in the full settings string. This change has the side effect and added benefit of requiring removal of the separator from the modifier names since the prepended separator will be checked when adding the modifier. No change will be required for the settings names themselves.