Prevent continuing setup after Miniconda installation on Linux
Ubuntu and likely other Linux distributions use a .bashrc script that prevents sourcing it non-interactively. After Miniconda installation, the bash profile/rc file needs to be re-sourced to find the conda executable, but it will not be found on these distributions. As a workaround, exit the setup script after Miniconda installation on all Linux platforms, informing the user to rerun the script in a new terminal.