Incorporate vmax saturation parameter into 3D surface preprocessing
Use the "clip_vmax" profile setting when saturating the image prior to surface viewing and decrease this value for the "atlas" profile modifier to allow greater saturation of atlas images. Tested on ABA P56, where this preprocessing value was necessary to avoid large gaps in the surface, and ABA E18pt5, where the saved file size was dramatically reduced.