No longer use label level -1 as a special case
Label level -1 was used as a special case to catch any blobs that did not fit into any subsequent regions, despite the hierarchical ontology organization that presupposed that any given level would be a subset of a higher level. This special case was used for new label IDs inadvertently generated during spline interpolation, which have been corrected previously. As a special case, this label continued to appear in volume dictionary output for each level. Removing this case avoids this redundancy. Note that the redundancy has been ignored while calculating stats in R since this region's volume is empty.