Change ellipsoid spacing, labels, and output array
When the spacing parameter is applied to the ellipsoid function, the resulting ellipsoid does not appear to be centered in the array, despite its API specification. To bypass this misplacement, simply multiply the axes by the spacing term instead. Also allow the spacing to be set as a paramter to allow customizing its value (or ignoring it), such as for microscopes where the effective spacing may differ from the resolution value.

- Use spacing as radius multiplier for each axis
- Allow spacing to be set as parameter
- Turn off spacing during blob segmentation so that greater z-axis coverage, though less so than before the spacing fix
- Set the output array as a parameter to allow specifying the data type
- Option to set the labels for each ellipsoid