Simplify recursive lateral edge extension
The edge extension function was called twice for each plane for each region, once to set up based on the current plane and the next to perform the resizing. To simplify, merge setup and resizing into the same call so that the function only needs to call itself recursively on the next plane.