Change bar width parameter to bar group padding
The ideal bar width is dependent on the number of bars in each group as well as the desired padding between groups. Rather than setting the absolute width of the bar, set the fraction of space that should be left unoccupied as padding and dynamically determine the bar width from the remaining space. Clarify the meaning of "group" in grouped bar plots as well. Also provide some horizontal padding space when tightening the layout to allow manually squeezing the resulting plot for small number of bars.