Fix label ID access during labels measurement with scalar label
Labels metric measurement had assumed that each label was a sequence, but they can be scalars as well. The labels metric class will now check whether the ID is a sequence and take only the first element for display and return purposes while still using the sequence to find the full region. This reduction simplifies the output to just the enclosing region both for display and the resulting data frame.