Sub-segment labels based on anatomical borders
Use atlas edges as a map of anatomical borders to guide sub-division of each label. Use a watershed with seeds and boundaries given by a distance transform to the anatomical borders, limiting the seeds and ensuring that sub-segments occupy at least 20% of original space. Fill in missing space with in-painting, with a commented-out section as a alternative approach where connected neighbors are iteratively merged into each retained sub-segment.