Use mask for labels watershed
The watershed function allows a mask parameter to avoid filling specified pixels, which can be used to specify the outer boundaries directly rather than through the distance transform there. Use of this mask obviates the need for write outer boundaries into the edge map or apply a mask after the watershed as well as influencing the flow to avoid any background areas.

- Option to apply mask for distance-based watershed
- Use a mask of the labels image if present without atlas image, opening the labels image slightly for smoother borders
- Change mask priorities when segmenting from labels, making the atlas image optional
- Decrease the compactness slightly to encourage staying within boundaries