Smooth mask for cropping watershed labels
Apply an opening filter to background mask used to crop the final labels. The labels will no longer match the farthest extent of the original labels, but this smoothing allows the labels to fill up some areas that may not have been covered by the original labels and potentially make up for gaps that they already had with the original labels.