Option to use only the atlas when generating background mask during atlas edge generation
Atlas edge generation uses a background mask to assist finding the farthest outside boundary using a combination of both the LoG of the atlas and the label image. Atlases that label the ventricular regions, however, do not fully benefit from this background mask in ventricular regions since the labels cover this border. To allow the background mask to include labeled ventricular regions, provide the new profile option, "log_atlas_thresh", to specify using the "atlas_threshold" value as a threshold for the atlas itself (not the LoG image), without the labels.

- Option to use atlas alone as background mask for atlas outside edge detection
- Employ this option for ABA E11pt5 along with a set atlas threshold