Add button to save 3D image as screenshot with transparency
Saving Mayavi-rendered 3D image using the Mayavi interface does not appear to allow transparency of the background. To save with transparency, provide a button next to the current save button to save 3D images specifically. This save will use the command-line given extension or fallback to png format. To save in 3D formats, the user can still supply a format such as .obj to save with the refresh button or via the Mayavi interface.