Remove Plot Editor zoom scaling by scaling factor
Zoom control modified x/y zoom speeds based on the scaling factor, but this scaling factor is often 1 for each dimension and caused an exception if not set. As scaling only describes the relationship between the image and a larger version of the image, and the resolution should already been incorporated into the displayed image, scaling does not appear to be necessary for modifying the zoom speed.