Fix "smooth" profile setting to accept floats
The label smoothing method checked for integers to determine whether the "smooth" setting indicated a single filter size or a sequence of sizes for testing multiple sizes, but these sizes can be floats, such as for the sigma in smoothing by Gaussian kernel. Check for sequence for test instead, assuming that all other values are for single filter sizes.