Option to duplicate positive labels as negative ones to keep the same color scheme after mirroring
Mirroring takes positive labels and converts them into negative ones for the opposite hemisphere. Adding these negative labels causes a shift in the discrete colormap screen so that corresponding labels in the original and mirrored labels have different colors. To map these labels to the same colors, provide the option to add colors for negative labels, even though colors for labels >= 0 will be used for the original labels image.

- Option to add colors for matching negative labels
- Employ this option in the AtlasEditor and stack export
- Avoid duplicate unique call
- Defer conversion to float until after duplications to minimize float recreation in intermediate arrays