Simplify conversion of filenames to Clrbrain-oriented paths
Clrbrain has accepted paths where the suffixes were appended directly to full filenames, after the extension, which required checking not for the last period but for the full extension. Now that we are using filenames with any suffixes spliced in before the extension instead to clarify potential confusion regarding extensions, simply split the filename at the extension rather than splicing out a specific extension.