Check for NaNs when generating labels difference images
NaNs typically occur when the measurement has not been recorded, meaning that the labels difference image will not show any information. Another scenario where NaNs appear are for regions that have not been measured, such as the background. While image can be saved with NaNs for these regions, application such as ITK-SNAP appear to be unable to handle these values and display all pixels as NaNs, so these regions are instead left as 0.

- Avoid generating difference image if measurement is not in data frame or only consists of NaNs
- Avoid outputting region difference if either of the first two conditions consists of NaNs, such as for background