Change default font families since DejaVuSans is not cross-platform by default
DejaVuSans is the new default for Matplotlib, which has a copy of this font so that it displays properly even on systems without the font. The font is also used for PDF export, however, meaning that it may display incorrectly on platforms without the font or in programs that attempt to import the PDF file, such as Inkscape. Change the default font to Arial since it is presumably present on most Mac and Windows platforms, and Matplotlib will still fallback to DejaVuSans if none of the other font options are available