Generalize image rotation after loading non-Clrbrain format to all main loaded images
Loading non-Clrbrain images applied the `--flip` parameter as a special case to allow for the 180 degree rotation required for another package. This rotation conflicted with stack export, which also utilized the parameter and re-flipped the image. To avoid this conflict and generalize rotation, use a separate `load_rot90` process setting to specify the number of 90deg rotations to make for the main image immediately after loading, regardless of image file type. The `--flip` parameter can then be used for on-the-fly flipping, such as before registration or image export.