Move Atlas Editor from 2D plot style to separate button
Make Atlas Editor more accessible and clarify that it is a separate plotter by creating a separate button for it. Also lift restriction on having at Atlas Editor and ROI Editor window open at the same time since the Atlas Editor save button is now separate from that of the ROI Editor.