Replace library method to get extension with splitting method support extensions with multiple periods
`get_filename_ext` essentially duplicated the functionality of `os.path.splitext` and was not widely used. Replace it with a method that expands upon the latter method by support extensions with multiple dots. This new method works under the assumption that certain extensions are merely the start of additional extensions, such as ".tar" for ".tar.gz" and ".nii" for ".nii.gz". Support these particular extension starts when splitting a path to split on the full extension rather than only the last component.