Fix to remove duplicate application of denominator for paired line segments in jitter plot
Simplify comparing paired values by storing only the last group rather than restoring the group in another list, which also avoid confusion from reuse of a variable name for a different purpose. This change also avoids the need to re-divide any of the groups by the denominator and fixes a duplication division.