Removing splitting ROI into stacks for nuclei detection
Now that intermediate denoised stacks are no longer saved, the ROI does not need to be split into stacks, or "super-ROIs," to fit into RAM. Instead, the number of sub-ROIs simply defined by the number of processors, which in turn determines RAM usage. By no longer splitting into stacks, the separate stack pruning step is no longer required, reducing the overall amount of pruning. The stack detector function remains a separate function but now for sub-ROIs to allow future processing of a given subset of sub-ROIs in distributed computing.

- Replace splitting into stacks with immediately spitting into sub-ROIs
- Remove merging blob pruning and detection time dataframes from each stack now that only one data frame each is required
- Display detection information for each sub-ROI even outside of verbose mode now that stack messages are no longer relevant