Avoid saving stack processing data frames when generating an ROC
ROCs may run through many iterations, saving data frames for blob pruning and stack detection times that may not be particularly relevant. Simply turn off this saving when building ROCs. Also support `None` as an option for exporting a list of data frames to CSV to allow calling the function to show even if no path is given, which can be useful to print the data frame.